LayoutTests/accessibility/mac/aria-tree-item-children-expected.txt

Animals
Birds
This tests that aria tree items can be fetched as either a full list of children or via range getter to ensure the range getter isn't returning a row as it's first child.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS tree.role is 'AXRole: AXOutline'
PASS child1.stringValue == child2.stringValue && child1.role == child2.role is true
PASS successfullyParsed is true

TEST COMPLETE
